refactor(hub): relocate adapter/store/util/query seams into kernel/ (stage 3e)
Move the four public-contract seam folders under src/kernel/ and point the
tsup entries at the new source paths. Output keys (adapter/index, store/index,
util/index, query/index) are unchanged, so the published dist subpaths
dist/{adapter,store,util,query}/index.js stay byte-identical. Moved files'
cross-out imports adjusted for the deeper location (../kernel/ -> ../,
../with-* -> ../../with-*). package.json exports untouched.
