+++ packages/hub/src/kernel/query/join.ts
@@ -45,10 +45,10 @@
  * including joined reads.
  */
 
-import type { RefDescriptor, RefMode } from '../kernel/refs.js'
+import type { RefDescriptor, RefMode } from '../refs.js'
 import { readPath } from './predicate.js'
-import { JoinTooLargeError, DanglingReferenceError } from '../kernel/errors.js'
-import { applyI18nLocale, type I18nTextDescriptor } from '../with-shape/i18n/core.js'
+import { JoinTooLargeError, DanglingReferenceError } from '../errors.js'
+import { applyI18nLocale, type I18nTextDescriptor } from '../../with-shape/i18n/core.js'
 
 /** Planner strategy for a single join leg. Auto-selected unless overridden. */
 export type JoinStrategy = 'hash' | 'nested'
