+++ packages/hub/src/kernel/util/discriminant.ts
