+++ packages/hub/src/kernel/util/sanitize-filename.ts
@@ -41,7 +41,7 @@
  * @module
  */
 
-import { FilenameSanitizationError } from '../kernel/errors.js'
+import { FilenameSanitizationError } from '../errors.js'
 
 /**
  * One of seven storage destinations the sanitizer knows how to defang
