+++ packages/hub/tsup.config.ts
@@ -21,13 +21,13 @@ import { defineConfig } from 'tsup'
 const ENTRIES = {
   index: 'src/index.ts',
   'i18n/index': 'src/with-shape/i18n/index.ts',
-  'store/index': 'src/store/index.ts',
+  'store/index': 'src/kernel/store/index.ts',
   'team/index': 'src/with-party/team/index.ts',
   'session/index': 'src/with-party/session/index.ts',
   'history/index': 'src/with-commit/history/index.ts',
   'forget/index': 'src/with-audit/forget/index.ts',
   'sealed-record/index': 'src/with-audit/sealed-record/index.ts',
-  'query/index': 'src/query/index.ts',
+  'query/index': 'src/kernel/query/index.ts',
   'blobs/index': 'src/with-shape/blobs/index.ts',
   'indexing/index': 'src/with-lookup/indexing/index.ts',
   'aggregate/index': 'src/with-lookup/aggregate/index.ts',
@@ -44,11 +44,11 @@ const ENTRIES = {
   'materialized-views/index': 'src/with-formula/materialized-views/index.ts',
   'overlay-views/index': 'src/with-formula/overlay-views/index.ts',
   'sync/index': 'src/with-party/sync/index.ts',
-  'util/index': 'src/util/index.ts',
+  'util/index': 'src/kernel/util/index.ts',
   'attestation/index': 'src/with-audit/attestation/index.ts',
   'kernel/index': 'src/kernel/index.ts',
   'cargo/index': 'src/with-cargo/index.ts',
-  'adapter/index': 'src/adapter/index.ts',
+  'adapter/index': 'src/kernel/adapter/index.ts',
   describe: 'src/describe.ts',
 }
